chore: weapon system wip: trigger behaviors

+++ Entities/Weapons/WeaponSystem/Interface/IAttackBehavior.cs
@@ -2,5 +2,5 @@
 
 public interface IAttackBehavior : IWeaponBehavior
 {
-    public bool ExecuteAttack(float force);
+    public bool ExecuteAttack(double force);
 }
